Refactor: Prepare codebase for Dashboard
