+++ src/app/(chat)/ask-anything/[chatId]/_components/ChatHeader.tsx
