+++ src/app/(chat)/ask-anything/[chatId]/_components/MessageContainer.tsx
@@ -3,7 +3,7 @@ import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter'
 import { oneDark } from 'react-syntax-highlighter/dist/esm/styles/prism'
 
 import { Message } from '@/app/_lib/db'
-import ShowMessage from '@/app/ask-anything/[chatId]/_components/ShowMessage'
+import ShowMessage from '@/app/(chat)/ask-anything/[chatId]/_components/ShowMessage'
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar'
 import { cn } from '@/lib/utils'
