+++ src/app/(chat)/ask-anything/[chatId]/_components/UploadDocumentWrapper.tsx
@@ -5,9 +5,9 @@ import { Tooltip } from '@nextui-org/tooltip'
 import { FileBarChart2, Paperclip } from 'lucide-react'
 
 import { Button } from '@/components/ui/button'
-import { Chat } from '../../../../app/_lib/db'
-import { truncateString } from '../../../../app/_utils'
-import { Badge } from '../../../../components/ui/badge'
+import { Chat } from '../../../../_lib/db'
+import { truncateString } from '../../../../_utils'
+import { Badge } from '../../../../../components/ui/badge'
 import { UploadDocument } from './UploadDocument'
 
 const UploadDocumentWrapper = ({
