+++ src/app/(chat)/ask-anything/[chatId]/actions/addDocument.ts
