+++ src/app/(chat)/ask-anything/[chatId]/utils/creatNewChat.ts
