+++ src/app/(chat)/ask-anything/[chatId]/utils/validateDocReq.ts
